Add files via upload
2. Define a class called  FourWheeler consisting of  name, cc, safety, price. 
If cc is less than 1000 then it will or greater than 4000, then it will generate VehicleException with information like ‘It is a Two Wheeler, not Four Wheeler’ or ‘It is a Vehicle not a Train ’ respectively. 
Similarly if the safety if not 5 then it’ll generate UnsafetyException with a message ‘Sorry, Invalid Safty features’. 
The VehicleException and UnsafetyExceptions are user defined exception classes. 
Create at least 4 object’s of FourWheer class in main() method of Factory class to validate the FourWheeres cc and safety.